+++ zcash_keys/CHANGELOG.md
@@ -13,6 +13,8 @@ workspace.
 - `zcash_keys::keys::zcashd`, a module containing utilities that are useful
   for interacting with key data from legacy zcashd wallets, has been added
   under a newly-added `zcashd-compat` feature flag.
+- `zcash_keys::address::UnifiedAddress::receivers`
+- `zcash_keys::address::Address::receivers`
 
 ## [0.9.0] - PLANNED